@@ -1,22 +1,23 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-  	<PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.0" />
-  	<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.0">
+  	<PackageReference Include="Microsoft.EntityFrameworkCore" Version="10.0.1" />
+  	<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="10.0.1">
   		<PrivateAssets>all</PrivateAssets>
   		<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
   	</PackageReference>
-  	<PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="9.0.0">
+  	<PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="10.0.1">
   		<PrivateAssets>all</PrivateAssets>
   		<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
   	</PackageReference>
-  	<PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.1" />
+  	<PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="10.0.0" />
+  	<PackageReference Include="Serilog.Sinks.Seq" Version="9.0.0" />
   </ItemGroup>
 	
   <ItemGroup>
