@@ -2,8 +2,8 @@
 
 public class JwtOption
 {
-    public string Key { get; set; }
-    public string Issuer { get; set; }
-    public string Audience { get; set; }
+    public string Key { get; set; } = null!;
+    public string Issuer { get; set; } = null!;
+    public string Audience { get; set; } = null!;
     public int ExpiresInMinutes { get; set; }
 }
